Don't store previous commit hash if cache was empty
This could end up badly if the previous build was a pull request
or a commit from another branch, which would not be existing in the
current branch and therefore always end up with a empty changelog.

By adding $DRONE_REPO_BRANCH this will make the .last_commit work
independently and store a local copy only if the log was not empty.